+++ Common/processfields_td.cpp
@@ -68,7 +68,7 @@ int ProcessFieldsTD::Process()
 		ss << std::setw( pad_length ) << std::setfill( '0' ) << m_Eng_Interface->GetNumberOfTimesteps();
 		size_t datasize[]={numLines[0],numLines[1],numLines[2]};
 		success &= m_HDF5_Dump_File->WriteVectorField(ss.str(), field, datasize);
-		float time[1]={m_Eng_Interface->GetTime(m_dualTime)};
+		float time[1] = {(float)m_Eng_Interface->GetTime(m_dualTime)};
 		success &= m_HDF5_Dump_File->WriteAtrribute("/FieldData/TD/"+ss.str(),"time",time,1);
 	}
 	else
